Melhorias de UI, estabilidade de saves, linkagem estática e polimento de sistemas
Compilação:
- Adicionadas flags de linkagem estática (-static, -static-libgcc, -static-libstdc++) no Makefile para gerar um .exe standalone (sem dependência de DLLs externas).

Interface e Visual (UI):
- Criada a nova função `gerarBarraGradiente` em Aparencia, substituindo a antiga `gerarBarraSuave`. Aplicada com sucesso na Tela de Atributos, Combate, HUD 3D e Tela de Vitória.
- Na tela de Level Up, os atributos agora mostram uma pré-visualização exata do ganho (ex: 150 -> 160 [+10]) com cores ANSI destacadas.
- Corrigido o bug visual na caixa de Drops do Mahoraga (e outros chefes), que cortava nomes de itens e quebrava o layout ASCII por não ignorar os bytes invisíveis das cores ANSI.

Inventário:
- O menu de equipamentos e consumíveis agora é dividido dinamicamente por categorias visuais ("EQUIPADOS", "ARMAS - DANO FÍSICO", "CURA E RESTAURAÇÃO", etc.).
- Melhoria nas mensagens de erro ao tentar usar itens inválidos, respondendo especificamente ao tipo (ex: avisando que Materiais devem ser levados a NPCs).
- Ajuste na nomenclatura das opções de interação para itens de Missão e Materiais.

Mapa Múndi (Fast Travel):
- Substituição das flags individuais de mapa por uma flag unificada `Mapas_Descobertos`.
- Ocultação dinâmica de áreas não visitadas, substituindo nomes por "???" e ajustando matematicamente as bordas ASCII da caixa para não deformar o desenho.

Saves e Progressão:
- Correção de BUG CRÍTICO de retrocompatibilidade: Jogadores com saves antigos agora conseguem carregar o jogo sem corromper a progressão (lógica segura para as variáveis novas de Parry, Regen do Troll e Ícones).
- Saves antigos recebem a nova flag de mapas retroativamente caso já tivessem visitado a Floresta ou Reino.
- O jogo agora salva corretamente o item equipado no Acesso Rápido (Consumível).

Combate e Gameplay:
- O minigame de Parry agora é desativado/ignorado automaticamente se o jogador utilizar a ação de "Defender" com um escudo no próprio turno.
- Tempo do Parry convertido para exibir apenas valores inteiros em segundos (removidas casas decimais soltas).
- Menu de Debug refatorado para usar switch-cases, unificando o desbloqueio de mapas e permitindo inserir a quantidade exata desejada de XP e Ouro.

+++ RPG C++/Sistemas/Combate/Parry.cpp
@@ -78,7 +78,7 @@ bool Parry::executarMinigame(int quantidadeDeNumerosParaDigitar, int tempoLimite
         auto tempoAtual = std::chrono::steady_clock::now();
         std::chrono::duration<double> tempoDecorrido = tempoAtual - tempoInicial;
         if (tempoDecorrido.count() > tempoLimiteEmSegundos) {
-            std::string timeoutMsg = "TEMPO ESGOTADO (" + std::to_string(tempoDecorrido.count()) + "s)!";
+            std::string timeoutMsg = "TEMPO ESGOTADO (" + std::to_string(static_cast<int>(tempoDecorrido.count())) + "s)!";
             std::cout << "\n" << Aparencia::margemCombate() << FuncoesDialogo::formatarMsgCombate(timeoutMsg, Cor::FUNDO_VERMELHO) << "\n";
             return false;
         }
